requirements.txt
asgiref==3.10.0
babel==2.17.0
Django==5.2.7
django-shortuuidfield==0.1.3
django-taggit==6.1.0
pillow==11.3.0
psycopg2-binary==2.9.11
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
